# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EV Electric Investment Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginRequest {
  phone: string;
  password: string;
}

export interface RegisterRequest {
  phone: string;
  password: string;
  name: string;
  referralCode?: string;
}

export interface User {
  id: number;
  name: string;
  phone: string;
  isAdmin: boolean;
  referralCode: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type UserWithBalance = User & {
  balance: number;
  totalInvested: number;
  totalEarnings: number;
};

export interface BalanceInfo {
  balance: number;
  totalInvested: number;
  totalEarnings: number;
  activeInvestments: number;
}

export interface DepositRequest {
  amount: number;
  paymentMethod: string;
  notes?: string;
}

export interface InvestmentPlan {
  id: number;
  name: string;
  description: string;
  minAmount: number;
  maxAmount: number;
  dailyReturnPercent: number;
  durationDays: number;
  isActive: boolean;
}

export interface CreateInvestmentPlanRequest {
  name: string;
  description: string;
  minAmount: number;
  maxAmount: number;
  dailyReturnPercent: number;
  durationDays: number;
  isActive?: boolean;
}

export interface UserInvestment {
  id: number;
  planId: number;
  planName: string;
  amount: number;
  dailyReturnPercent: number;
  durationDays: number;
  status: string;
  startDate: string;
  endDate: string;
  totalEarned: number;
}

export interface CreateInvestmentRequest {
  planId: number;
  amount: number;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  deposit: "deposit",
  withdrawal: "withdrawal",
  investment: "investment",
  earning: "earning",
} as const;

export type TransactionStatus =
  (typeof TransactionStatus)[keyof typeof TransactionStatus];

export const TransactionStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Transaction {
  id: number;
  userId: number;
  userName: string;
  type: TransactionType;
  amount: number;
  status: TransactionStatus;
  paymentMethod?: string;
  notes?: string;
  createdAt: string;
}

export interface UpdateBalanceRequest {
  balance: number;
  note?: string;
}