# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: EV Electric Investment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/deposit:
    post:
      operationId: deposit
      tags: [users]
      summary: Request a deposit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositRequest"
      responses:
        "200":
          description: Deposit request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
...
components:
  schemas:
...
    DepositRequest:
      type: object
      properties:
        amount:
          type: number
        paymentMethod:
          type: string
        notes:
          type: string
      required:
        - amount
        - paymentMethod
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        type:
          type: string
          enum: [deposit, withdrawal, investment, earning]
        amount:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected]
        paymentMethod:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - userName
        - type
        - amount
        - status
        - createdAt
...